id type changed

+++ src/main/java/org/jugistanbul/dto/UserDTO.java
@@ -4,5 +4,5 @@ package org.jugistanbul.dto;
  * @author hakdogan (huseyin.akdogan@patikaglobal.com)
  * Created on 20.10.2021
  */
-public record UserDTO(Integer id, String username, String password, String role) {
+public record UserDTO(Long id, String username, String password, String role) {
 }
